requirements.txt
absl-py==1.4.0
attrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
contourpy==1.0.7
cvzone==1.5.6
cycler==0.11.0
decorator==4.4.2
exceptiongroup==1.1.0
fonttools==4.38.0
future==0.18.3
huepy==1.2.1
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
importlib-resources==5.12.0
iniconfig==2.0.0
instabot==0.117.0
kiwisolver==1.4.4
matplotlib==3.7.0
mediapipe==0.8.9.1
mock==5.0.1
moviepy==1.0.3
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
pilgram==1.2.1
Pillow==9.4.0
pluggy==1.0.0
proglog==0.1.10
protobuf==3.19.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.22.0
schedule==1.1.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
tweepy==4.12.1
types-toml==0.10.8.5
urllib3==1.26.14
Wand==0.6.11
zipp==3.14.0
